// sdks/typescript/src/apache_beam/proto/metrics.ts
export interface MonitoringInfo {
    /**
     * (Required) Defines the semantic meaning of the metric or monitored state.
     *
     * See MonitoringInfoSpecs.Enum for the set of well known metrics/monitored
     * state.
     *
     * @generated from protobuf field: string urn = 1;
     */
    urn: string;
    /**
     * (Required) Defines the encoding and aggregation method for the payload.
     *
     * See MonitoringInfoTypeUrns.Enum for the set of well known types.
     *
     * @generated from protobuf field: string type = 2;
     */
    type: string;
    /**
     * (Required) The metric or monitored state encoded as per the specification
     * defined by the type.
     *
     * @generated from protobuf field: bytes payload = 3;
     */
    payload: Uint8Array;
    /**
     * A set of key and value labels which define the scope of the metric. For
     * well known URNs, the set of required labels is provided by the associated
     * MonitoringInfoSpec.
     *
     * Either a well defined entity id for matching the enum names in
     * the MonitoringInfoLabels enum or any arbitrary label
     * set by a custom metric or user metric.
     *
     * A monitoring system is expected to be able to aggregate the metrics
     * together for all updates having the same URN and labels. Some systems such
     * as Stackdriver will be able to aggregate the metrics using a subset of the
     * provided labels
     *
     * @generated from protobuf field: map<string, string> labels = 4;
     */
    labels: {
        [key: string]: string;
    };
    /**
     * This indicates the start of the time range over which this value was
     * measured.
     * This is needed by some external metric aggregation services
     * to indicate when the reporter of the metric first began collecting the
     * cumulative value for the timeseries.
     * If the SDK Harness restarts, it should reset the start_time, and reset
     * the collection of cumulative metrics (i.e. start to count again from 0).
     * HarnessMonitoringInfos should set this start_time once, when the
     * MonitoringInfo is first reported.
     * ProcessBundle MonitoringInfos should set a start_time for each bundle.
     *
     * @generated from protobuf field: google.protobuf.Timestamp start_time = 5;
     */
    startTime?: Timestamp;
}
// ...
class MonitoringInfo$Type extends MessageType<MonitoringInfo> {
    constructor() {
        super("org.apache.beam.model.pipeline.v1.MonitoringInfo", [
            { no: 1, name: "urn", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "type", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "payload", kind: "scalar", T: 12 /*ScalarType.BYTES*/ },
            { no: 4, name: "labels", kind: "map", K: 9 /*ScalarType.STRING*/, V: { kind: "scalar", T: 9 /*ScalarType.STRING*/ } },
            { no: 5, name: "start_time", kind: "message", T: () => Timestamp }
        ]);
    }
    create(value?: PartialMessage<MonitoringInfo>): MonitoringInfo {
        const message = { urn: "", type: "", payload: new Uint8Array(0), labels: {} };
        globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
        if (value !== undefined)
            reflectionMergePartial<MonitoringInfo>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: MonitoringInfo): MonitoringInfo {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string urn */ 1:
                    message.urn = reader.string();
                    break;
                case /* string type */ 2:
                    message.type = reader.string();
                    break;
                case /* bytes payload */ 3:
                    message.payload = reader.bytes();
                    break;
                case /* map<string, string> labels */ 4:
                    this.binaryReadMap4(message.labels, reader, options);
                    break;
                case /* google.protobuf.Timestamp start_time */ 5:
                    message.startTime = Timestamp.internalBinaryRead(reader, reader.uint32(), options, message.startTime);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    private binaryReadMap4(map: MonitoringInfo["labels"], reader: IBinaryReader, options: BinaryReadOptions): void {
        let len = reader.uint32(), end = reader.pos + len, key: keyof MonitoringInfo["labels"] | undefined, val: MonitoringInfo["labels"][any] | undefined;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case 1:
                    key = reader.string();
                    break;
                case 2:
                    val = reader.string();
                    break;
                default: throw new globalThis.Error("unknown map entry field for field org.apache.beam.model.pipeline.v1.MonitoringInfo.labels");
            }
        }
        map[key ?? ""] = val ?? "";
    }
    internalBinaryWrite(message: MonitoringInfo, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string urn = 1; */
        if (message.urn !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.urn);
        /* string type = 2; */
        if (message.type !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.type);
        /* bytes payload = 3; */
        if (message.payload.length)
            writer.tag(3, WireType.LengthDelimited).bytes(message.payload);
        /* map<string, string> labels = 4; */
        for (let k of Object.keys(message.labels))
            writer.tag(4, WireType.LengthDelimited).fork().tag(1, WireType.LengthDelimited).string(k).tag(2, WireType.LengthDelimited).string(message.labels[k]).join();
        /* google.protobuf.Timestamp start_time = 5; */
        if (message.startTime)
            Timestamp.internalBinaryWrite(message.startTime, writer.tag(5, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message org.apache.beam.model.pipeline.v1.MonitoringInfo
 */
export const MonitoringInfo = new MonitoringInfo$Type();